% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-docs.R
\name{anova}
\alias{anova}
\alias{anova.wls}
\alias{anova.meta}
\alias{anova.meta3LFIML}
\alias{anova.reml}
\alias{anova.osmasem}
\alias{anova.osmasem2}
\alias{anova.mxsem}
\title{Compare Nested Models with Likelihood Ratio Statistic}
\arguments{
\item{object}{An object or a list of objects of various classes. It will be
passed to the \code{base} argument in \code{\link[OpenMx]{mxCompare}}.}

\item{\dots}{An object or a list of objects of various classes. It will be
passed to the \code{comparison} argument in \code{\link[OpenMx]{mxCompare}}.}

\item{all}{A Boolean value on whether to compare all bases with all
comparisons. It will be passed to the \code{all} argument in
\code{\link[OpenMx]{mxCompare}}.}
}
\value{
A table of comparisons between the models in base and comparison.
}
\description{
It compares nested models with the likelihood ratio statistic from various
objects. It is a wrapper of \code{\link[OpenMx]{mxCompare}}.
}
\note{
When the objects are class \code{\link[metaSEM]{wls}}, the degrees of
freedom in the base and comparison models are incorrect, while the degrees
of freedom of the difference between them is correct. If users want to
obtain the correct degrees of freedom in the base and comparison models,
they may individually apply the \code{\link[metaSEM]{summary}} function on
the base and comparison models.
}
\examples{

## Test the significance of a predictor with likelihood ratio test
## Model0: No predictor
model0 <- meta(y=yi, v=vi, data=Hox02, model.name="No predictor")

## Model1: With a predictor
model1 <- meta(y=yi, v=vi, x=weeks, data=Hox02, model.name="One predictor")

## Compare these two models
anova(model1, model0) 

}
\author{
Mike W.-L. Cheung <mikewlcheung@nus.edu.sg>
}
\keyword{methods}
