% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-docs.R
\name{coef}
\alias{coef}
\alias{coef.tssem1FEM}
\alias{coef.tssem1FEM.cluster}
\alias{coef.tssem1REM}
\alias{coef.wls}
\alias{coef.wls.cluster}
\alias{coef.meta}
\alias{coef.meta3LFIML}
\alias{coef.reml}
\alias{coef.osmasem}
\alias{coef.osmasem2}
\alias{coef.mxsem}
\title{Extract Parameter Estimates from various classes.}
\arguments{
\item{object}{An object returned from either class \code{tssem1FEM}, class
\code{tssem1FEM.cluster}, class \code{tssem1REM}, class \code{wls}, class
\code{wls.cluster}, class \code{meta}, class \code{reml}, class
\code{osmasem}, class \code{osmasem2}, or class \code{sem}}

\item{select}{Select \code{all} for both fixed- and random-effects
parameters, \code{fixed} for the fixed-effects parameters or \code{random}
for the random-effects parameters. For \code{meta3LFIML} objects,
\code{allX} is used to extract all parameters including the predictors and
auxiliary variables.}

\item{\dots}{Further arguments; currently none is used}
}
\value{
Parameter estimates for both fixed-effects (if any) and
random-effects (if any)
}
\description{
It extracts the parameter estimates from objects of various classes.
}
\note{
\code{coef.sem} is simply a wraper of \code{omxGetParameters}. Extra
arguments will be passed to it
}
\examples{

## Random-effects meta-analysis
model1 <- meta(y=yi, v=vi, data=Hox02)
coef(model1)

## Fixed-effects only
coef(model1, select="fixed")

}
\seealso{
\code{\link[metaSEM]{tssem1}}, \code{\link[metaSEM]{wls}},
\code{\link[metaSEM]{meta}}, \code{\link[metaSEM]{reml}},
\code{\link[OpenMx]{omxGetParameters}}, \code{\link[metaSEM]{osmasem}}
}
\author{
Mike W.-L. Cheung <mikewlcheung@nus.edu.sg>
}
\keyword{methods}
