\name{Craft03}
\alias{Craft03}
\docType{data}
\title{Ten Correlation Matrices from Craft et al. (2003)
}
\description{
  This data set includes ten studies on the relationships between CSAI
  subscales and sport behavior. The original data were used in Craft et
  al. (2003), whereas a subset of them were illustrated in Becker (2009).
}
\usage{data(Craft03)}

\details{
A list of data with the following structure:
\describe{
  \item{data}{A list of 10 correlation matrices. The
	variables are \emph{Performance}, \emph{Cognitive}, \emph{Somatic}
	and \emph{Self confidence}}
  \item{n}{A vector of sample sizes}
  \item{Type_of_sport}{\emph{Individual} or \emph{Team}}
}
}

\source{
Craft, L. L., Magyar, T. M., Becker, B. J., & Feltz, D. L. (2003). The relationship between the Competitive State Anxiety Inventory-2 and sport performance: a meta-analysis. \emph{Journal of Sport and Exercise Psychology}, \bold{25(1)}, 44-65.
}
\references{
  Becker, B. J. (2009). Model-based meta-analysis. In H. Cooper,
  L. V. Hedges, & J. C. Valentine (Eds.), \emph{The Handbook of Research
  Synthesis and Meta-analysis} (2nd ed., pp. 377-395). New York: Russell
  Sage Foundation.
}
\examples{
data(Craft03)

}
\keyword{datasets}
