% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmasem.R
\name{create.V}
\alias{create.V}
\title{Create a V-known matrix}
\usage{
create.V(x, type = c("Symm", "Diag", "Full"), as.mxMatrix = TRUE)
}
\arguments{
\item{x}{A character vector of variable names of the sampling covariance
matrix.}

\item{type}{Either \code{"Symm"}, \code{"Diag"} or \code{"Full"}. Suppose
the number of variables is \eqn{p}, the numbers of variable names for
\code{"Symm"}, \code{"Diag"}, and \code{"Full"} are \eqn{p(p-1)/2
}{p(p-1)/2}, \eqn{p}{p}, and \eqn{p*p}{p*p}, respectively. The elements are
arranged in a column major.}

\item{as.mxMatrix}{Logical. Whether to convert the output into
\code{MxMatrix-class}.}
}
\value{
A list of \code{MxMatrix-class}. The V-known sampling covariance
matrix is computed in \code{V}.
}
\description{
It creates a V-known matrix of the sampling covariance matrix using
definition variables.
}
\examples{

\donttest{
my.df <- Cor2DataFrame(Nohe15A1)

## Create known sampling variance covariance matrix
V0 <- create.V(my.df$vlabels)
V0
}

}
\seealso{
\code{\link[metaSEM]{osmasem}}, \code{\link[metaSEM]{create.Tau2}},
\code{\link[metaSEM]{create.vechsR}}
}
\author{
Mike W.-L. Cheung <mikewlcheung@nus.edu.sg>
}
\keyword{osmasem}
