% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-docs.R
\docType{data}
\name{HedgesOlkin85}
\alias{HedgesOlkin85}
\title{Effects of Open Education Reported by Hedges and Olkin (1985)}
\source{
Hedges, L. V., & Olkin, I. (1985). \emph{Statistical methods for
meta-analysis.} Orlando, FL: Academic Press.
}
\description{
Effects of open education on attitude toward school and on reading
achievement reported by Hedges and Olkin (1985).
}
\details{
The variables are: \describe{ \item{study}{Study number}
\item{d_att}{Standardized mean difference on \emph{attitude}}
\item{d_ach}{Standardized mean difference on \emph{achievement}}
\item{var_att}{Sampling variance of the effect size of \emph{attitude}}
\item{cov_att_ach}{Sampling covariance between the effect sizes}
\item{var_ach}{Sampling variance of the effect size of \emph{achievement}} }
}
\examples{

data(HedgesOlkin85)

## Fixed-effects meta-analysis
summary( meta(y=cbind(d_att, d_ach),
              v=cbind(var_att, cov_att_ach, var_ach),
              data=HedgesOlkin85,
              RE.constraints=matrix(0, nrow=2, ncol=2)) )

}
\references{
Cheung, M. W.-L. (2010). Fixed-effects meta-analyses as
multiple-group structural equation models. \emph{Structural Equation
Modeling}, \bold{17}, 481-509.
}
\keyword{datasets}
