% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmasem.R
\name{osmasemSRMR}
\alias{osmasemSRMR}
\title{Calculate the SRMR in OSMASEM and OSMASEM3L}
\usage{
osmasemSRMR(x)
}
\arguments{
\item{x}{An OSMASEM object without any moderators.}
}
\value{
It calculates the model implied correlation matrix and its saturated
counterpart to calculate the SRMR. It should be noted that the heterogeneity
variances are ignored in the calculations.
}
\description{
It calculates the standardized root mean squared residuals (SRMR) in OSMASEM
and OSMASEM3L.
}
\seealso{
\code{\link[metaSEM]{osmasem}}, \code{\link[metaSEM]{Nohe15}}
}
\author{
Mike W.-L. Cheung <mikewlcheung@nus.edu.sg>
}
\keyword{osmasem}
\keyword{osmasem3L}
