% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-docs.R
\name{print}
\alias{print}
\alias{print.tssem1FEM}
\alias{print.tssem1FEM.cluster}
\alias{print.tssem1REM}
\alias{print.wls}
\alias{print.meta}
\alias{print.meta3LFIML}
\alias{print.reml}
\alias{print.uniR1}
\alias{print.impliedR}
\title{Print Methods for various Objects}
\arguments{
\item{x}{An object returned from either class \code{tssem1FEM}, class
\code{tssem1FEM.cluster}, class \code{tssem1REM}, class \code{wls}, class
\code{meta}, class \code{meta3LFIML}, class \code{reml}, class \code{uniR1}
or class \code{impliedR}}

\item{\dots}{Further arguments to be passed to \code{summary.default} or
unused.}
}
\description{
Print methods for the \code{tssem1FEM}, \code{tssem1FEM.cluster},
\code{tssem1REM}, \code{wls}, \code{meta}, \code{meta3LFIML}, \code{reml},
\code{uniR1} and \code{impliedR} objects.
}
\seealso{
\code{\link[metaSEM]{tssem1}}, \code{\link[metaSEM]{wls}},
\code{\link[metaSEM]{meta}}, \code{\link[metaSEM]{reml}}
}
\author{
Mike W.-L. Cheung <mikewlcheung@nus.edu.sg>
}
\keyword{methods}
