% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-docs.R
\docType{data}
\name{Roorda11}
\alias{Roorda11}
\title{Studies on Students' School Engagement and Achievement Reported by Roorda et
al. (2011)}
\source{
Roorda, D. L., Koomen, H. M. Y., Spilt, J. L., & Oort, F. J. (2011).
The influence of affective teacher-student relationships on students' school
engagement and achievement a meta-analytic approach. \emph{Review of
Educational Research}, \emph{81}(4), 493-529.
}
\description{
The data set includes 45 studies on the influence of affective
teacher-student relationships on students' school engagement and achievement
reported by Roorda et al. (2011).
}
\details{
The variables are: \describe{ \item{data}{A list of 45 studies of
correlation matrices. The variables are \emph{pos} (positive teacher-student
relations), \emph{neg} (negative teacher-student relations), \emph{enga}
(student engagement), and \emph{achiev} (student achievement).} \item{n}{A
vector of sample sizes} \item{SES}{A vector of average socio-economic status
(SES) of the samples} }
}
\examples{

\donttest{

## Random-effects model: First stage analysis
random1 <- tssem1(Cov = Roorda11$data, n = Roorda11$n, method = "REM",
                  RE.type = "Diag")
summary(random1)

## Model specification in lavaan syntax
model <- "## Regression paths
          enga ~ b31*pos + b32*neg
          achiev ~ b43*enga
          ## Variances of pos and neg are fixed at 1
          pos ~~ 1*pos
          neg ~~ 1*neg
          ## Correlation between pos and neg
          pos ~~ p21*neg
          ## Error variances
          enga ~~ p33*enga
          achiev ~~ p44*achiev"

RAM <- lavaan2RAM(model, obs.variables=c("pos", "neg", "enga", "achiev"))
RAM

## ## Equivalent RAM specification using create.mxMatrix()
## varnames <- c("pos", "neg", "enga", "achiev")
## A <- create.mxMatrix(c(0,0,0,0,
##                        0,0,0,0,
##                        "0.1*b31","0.1*b32",0,0,
##                        0,0,"0.1*b43",0),
##                      type = "Full", nrow = 4, ncol = 4, byrow = TRUE,
##                      name = "A", as.mxMatrix = FALSE)
## dimnames(A) <- list(varnames, varnames)
## A
##
## S <- create.mxMatrix(c(1,
##                        ".5*p21",1,
##                        0,0,"0.6*p33",
##                        0,0,0,"0.6*p44"),
##                      type="Symm", byrow = TRUE,
##                      name="S", as.mxMatrix = FALSE)
## dimnames(S) <- list(varnames, varnames)
## S

## Random-effects model: Second stage analysis
random2 <- tssem2(random1, RAM=RAM, diag.constraints=TRUE,
                  intervals="LB")
summary(random2)

## Display the model with the parameter estimates    
plot(random2)
}

}
\references{
Jak, S., & Cheung, M. W.-L. (2018). Addressing heterogeneity in
meta-analytic structural equation modeling using subgroup analysis.
\emph{Behavior Research Methods}, \bold{50}, 1359-1373.
}
\keyword{datasets}
