% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-docs.R
\name{summary}
\alias{summary}
\alias{summary.tssem1FEM}
\alias{summary.tssem1FEM.cluster}
\alias{summary.tssem1REM}
\alias{summary.wls}
\alias{summary.wls.cluster}
\alias{summary.meta}
\alias{summary.meta3LFIML}
\alias{summary.reml}
\alias{summary.CorPop}
\alias{summary.Cor3L}
\alias{summary.bootuniR2}
\alias{summary.osmasem}
\alias{summary.osmasem2}
\alias{summary.mxsem}
\alias{print.summary.tssem1FEM}
\alias{print.summary.tssem1FEM.cluster}
\alias{print.summary.wls}
\alias{print.summary.meta}
\alias{print.summary.meta3LFIML}
\alias{print.summary.reml}
\alias{print.summary.CorPop}
\alias{print.summary.Cor3L}
\alias{print.summary.bootuniR2}
\alias{print.summary.mxsem}
\title{Summary Method for tssem1, wls, meta, and meta3LFIML Objects}
\arguments{
\item{object}{An object returned from either class \code{tssem1FEM}, class
\code{tssem1FEM.cluster}, class \code{tssem1REM}, class \code{wls}, class
\code{wls.cluster}, class \code{meta}, class \code{meta3LFIML}, class
\code{reml}, class \code{mxsem} or class \code{CorPop}.}

\item{x}{An object returned from either class \code{summary.tssem1FEM},
class \code{tssem1FEM.cluster}, class \code{summary.wls}, class
\code{summary.meta}, class \code{summary.meta3LFIML}, class
\code{summary.reml} or class \code{summary.CorPop}.}

\item{homoStat}{Logical. Whether to conduct a homogeneity test on the effect
sizes.}

\item{allX}{Logical. Whether to report the predictors and the auxiliary
variables.}

\item{robust}{Logicial. Whether to use robust standard error from
\code{\link[OpenMx]{imxRobustSE}}.}

\item{df.adjustment}{Numeric. Adjust the degrees of freedom manually. It may
be necessary if the df calculated is incorrect when
\code{diag.constraints=TRUE}.}

\item{probs}{Quantiles for the parameter estimates.}

\item{cutoff.chisq.pvalue}{Cutoff of the p-value for the chi-square
statistic.}

\item{cutoff.CFI}{The cutoff of the CFI.}

\item{cutoff.SRMR}{The cutoff of the SRMR.}

\item{cutoff.RMSEA}{The cutoff of the RMSEA.}

\item{fitIndices}{Whether to calculate the chi-square statistic and various
goodness-of-fit indices in osmasem. Note. It may take a while since
statistics of the saturated and independence models are required.}

\item{numObs}{The number of observations in calculating the fit statistics
in osmasem. If it is missing, the total number of observations is used.}

\item{\dots}{Further arguments to be passed to
\code{\link[stats]{printCoefmat}}}
}
\description{
It summaries results for various class.
}
\note{
If the OpenMx status1 is either 0 or 1, the estimation is considered
fine. If the OpenMx status1 is other values, it indicates estimation
problems. Users should refer to `OpenMx` website for more details.
}
\seealso{
\code{\link[metaSEM]{tssem1}}, \code{\link[metaSEM]{wls}},
\code{\link[metaSEM]{meta}}, \code{\link[metaSEM]{reml}},
\code{\link[metaSEM]{rCor}}, \code{\link[metaSEM]{bootuniR2}},
\code{\link[metaSEM]{osmasem}}
}
\author{
Mike W.-L. Cheung <mikewlcheung@nus.edu.sg>
}
\keyword{methods}
