% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmasem.R
\name{VarCorr}
\alias{VarCorr}
\title{Extract Variance-Covariance Matrix of the Random Effects}
\usage{
VarCorr(x, ...)
}
\arguments{
\item{x}{An object returned from either class \code{meta} or \code{osmasem}}

\item{\dots}{Further arguments; currently none is used}
}
\value{
A variance-covariance matrix of the random effects.
}
\description{
It extracts the variance-covariance matrix of the random effects (variance
component) from either the \code{meta} or \code{osmasem} objects.
}
\note{
It is similar to \code{coef(object, select="random")} in tssem. The
main difference is that \code{coef()} returns a vector while
\code{VarCorr()} returns its correspondent matrix.
}
\examples{

## Multivariate meta-analysis on the log of the odds
## The conditional sampling covariance is 0
bcg <- meta(y=cbind(ln_Odd_V, ln_Odd_NV), data=BCG,
            v=cbind(v_ln_Odd_V, cov_V_NV, v_ln_Odd_NV))
VarCorr(bcg)

}
\seealso{
\code{\link[metaSEM]{coef}}, \code{\link[metaSEM]{vcov}}
}
\author{
Mike W.-L. Cheung <mikewlcheung@nus.edu.sg>
}
\keyword{methods}
