% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-docs.R
\name{vcov}
\alias{vcov}
\alias{vcov.tssem1FEM}
\alias{vcov.tssem1FEM.cluster}
\alias{vcov.tssem1REM}
\alias{vcov.wls}
\alias{vcov.wls.cluster}
\alias{vcov.meta}
\alias{vcov.meta3LFIML}
\alias{vcov.reml}
\alias{vcov.osmasem}
\alias{vcov.osmasem2}
\alias{vcov.mxsem}
\title{Extract Covariance Matrix Parameter Estimates from Objects of Various
Classes}
\arguments{
\item{object}{An object returned from objects of various classes}

\item{select}{Select \code{all} for both fixed- and random-effects
parameters, \code{fixed} for the fixed-effects parameters or \code{random}
for the random-effects parameters. For \code{meta3LFIML} objects,
\code{allX} is used to extract all parameters including the predictors and
auxiliary variables.}

\item{robust}{Logicial. Whether to use robust standard error from
\code{\link[OpenMx]{imxRobustSE}}.}

\item{\dots}{Further arguments; currently not in use except for
\code{tssemRobust1}, which to be passed to \code{\link[metafor]{robust}}.}
}
\value{
A variance-covariance matrix of the parameter estimates.
}
\description{
It extracts the variance-covariance matrix of the parameter estimates from
objects of various classes.
}
\note{
\code{vcov} returns \code{NA} when the \code{diag.constraints=TRUE}
argument is used in \code{wls} objects.
}
\examples{

## Random-effects meta-analysis
model1 <- meta(y=yi, v=vi, data=Hox02)
vcov(model1)

## Fixed-effects only
vcov(model1, select="fixed")

## Random-effects only
vcov(model1, select="random")

}
\seealso{
\code{\link[metaSEM]{tssem1}}, \code{\link[metaSEM]{wls}},
\code{\link[metaSEM]{meta}}, \code{\link[metaSEM]{reml}}
}
\author{
Mike W.-L. Cheung <mikewlcheung@nus.edu.sg>
}
\keyword{methods}
