% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec2symMat.R
\name{vec2symMat}
\alias{vec2symMat}
\title{Convert a Vector into a Symmetric Matrix}
\usage{
vec2symMat(x, diag = TRUE, byrow = FALSE)
}
\arguments{
\item{x}{A vector of numerics or characters}

\item{diag}{Logical. If it is \code{TRUE} (the default), the diagonals of
the created matrix are replaced by elements of \code{x}; otherwise, the
diagonals of the created matrix are replaced by "1".}

\item{byrow}{Logical. If it is \code{FALSE} (the default), the created
matrix is filled by columns; otherwise, the matrix is filled by rows.}
}
\value{
A symmetric square matrix.
}
\description{
It converts a vector into a symmetric matrix by filling up the elements into
the lower triangle of the matrix.
}
\examples{

vec2symMat(1:6)
#      [,1] [,2] [,3]
# [1,]    1    2    3
# [2,]    2    4    5
# [3,]    3    5    6

vec2symMat(1:6, diag=FALSE)
#      [,1] [,2] [,3] [,4]
# [1,]    1    1    2    3
# [2,]    1    1    4    5
# [3,]    2    4    1    6
# [4,]    3    5    6    1

vec2symMat(letters[1:6])
#      [,1] [,2] [,3]
# [1,] "a"  "b"  "c" 
# [2,] "b"  "d"  "e" 
# [3,] "c"  "e"  "f" 

}
\seealso{
\code{\link[metaSEM]{matrix2bdiag}}
}
\author{
Mike W.-L. Cheung <mikewlcheung@nus.edu.sg>
}
\keyword{utilities}
