\name{Digman97}
\alias{Digman97}
\alias{Digman97.n}
\docType{data}
\title{Factor correlation matrices of the Big Five Model from Digman (1997)}
\description{
Fourteen studies of the factor correlation matrices of the Big Five Model of personality reported by Digman (1997). 
}
\usage{data(Digman97)}
\format{
  The format is:
List of 14
 $ Digman 1 (1994)                : num [1:5, 1:5] 1 -0.48 -0.1 0.27 0.37 -0.48 1 0.62 0.41 0 ...
  ..- attr(*, "dimnames")=List of 2
  .. ..$ : chr [1:5] "E" "A" "C" "ES" ...
  .. ..$ : chr [1:5] "E" "A" "C" "ES" ...
 $ Digman 2 (1994)                : num [1:5, 1:5] 1 -0.3 0.07 0.09 0.45 -0.3 1 0.39 0.53 -0.05 ...
  ..- attr(*, "dimnames")=List of 2
  .. ..$ : chr [1:5] "E" "A" "C" "ES" ...
  .. ..$ : chr [1:5] "E" "A" "C" "ES" ...
 $ Digman 3 (1963c)               : num [1:5, 1:5] 1 0.25 -0.1 0.24 0.41 0.25 1 0.65 0.35 0.14 ...
  ..- attr(*, "dimnames")=List of 2
  .. ..$ : chr [1:5] "E" "A" "C" "ES" ...
  .. ..$ : chr [1:5] "E" "A" "C" "ES" ...
 $ Digman & Takemoto-Chock (1981b): num [1:5, 1:5] 1 -0.26 -0.16 0.01 0.66 -0.26 1 0.65 0.7 -0.03 ...
  ..- attr(*, "dimnames")=List of 2
  .. ..$ : chr [1:5] "E" "A" "C" "ES" ...
  .. ..$ : chr [1:5] "E" "A" "C" "ES" ...
 $ Graziano & Ward (1992)         : num [1:5, 1:5] 1 0.29 0.16 0.32 0.53 0.29 1 0.64 0.35 0.22 ...
  ..- attr(*, "dimnames")=List of 2
  .. ..$ : chr [1:5] "E" "A" "C" "ES" ...
  .. ..$ : chr [1:5] "E" "A" "C" "ES" ...
 $ Yik & Bond (1993)              : num [1:5, 1:5] 1 0.35 0.2 0.49 0.59 0.35 1 0.66 0.57 0.38 ...
  ..- attr(*, "dimnames")=List of 2
  .. ..$ : chr [1:5] "E" "A" "C" "ES" ...
  .. ..$ : chr [1:5] "E" "A" "C" "ES" ...
 $ John et al. 1 (1984)           : num [1:5, 1:5] 1 0.13 0.43 0.37 0.35 0.13 1 0.25 0.59 0.15 ...
  ..- attr(*, "dimnames")=List of 2
  .. ..$ : chr [1:5] "E" "A" "C" "ES" ...
  .. ..$ : chr [1:5] "E" "A" "C" "ES" ...
 $ John et al. 2 (1984)           : num [1:5, 1:5] 1 0.16 0.26 0.36 0.33 0.16 1 0.36 0.41 0.19 ...
  ..- attr(*, "dimnames")=List of 2
  .. ..$ : chr [1:5] "E" "A" "C" "ES" ...
  .. ..$ : chr [1:5] "E" "A" "C" "ES" ...
 $ Costa & McCrae 1 (1992c)       : num [1:5, 1:5] 1 0.11 0.19 0.22 0.56 0.11 1 0.18 0.44 0.24 ...
  ..- attr(*, "dimnames")=List of 2
  .. ..$ : chr [1:5] "E" "A" "C" "ES" ...
  .. ..$ : chr [1:5] "E" "A" "C" "ES" ...
 $ Costa & McCrae 2 (1992b)       : num [1:5, 1:5] 1 0.42 0.25 0.26 0.46 0.42 1 0.34 0.69 0.44 ...
  ..- attr(*, "dimnames")=List of 2
  .. ..$ : chr [1:5] "E" "A" "C" "ES" ...
  .. ..$ : chr [1:5] "E" "A" "C" "ES" ...
 $ Costa & McCrae 3 (1992b)       : num [1:5, 1:5] 1 0.04 0.27 0.21 0.4 0.04 1 0.24 0.25 -0.02 ...
  ..- attr(*, "dimnames")=List of 2
  .. ..$ : chr [1:5] "E" "A" "C" "ES" ...
  .. ..$ : chr [1:5] "E" "A" "C" "ES" ...
 $ Costa, McCrae, & Dye (1991)    : num [1:5, 1:5] 1 -0.07 0.22 0.21 0.43 -0.07 1 0.13 0.25 -0.06 ...
  ..- attr(*, "dimnames")=List of 2
  .. ..$ : chr [1:5] "E" "A" "C" "ES" ...
  .. ..$ : chr [1:5] "E" "A" "C" "ES" ...
 $ Barrick & Mount (1993)         : num [1:5, 1:5] 1 -0.04 -0.03 -0.03 0.28 -0.04 1 0.25 0.34 -0.17 ...
  ..- attr(*, "dimnames")=List of 2
  .. ..$ : chr [1:5] "E" "A" "C" "ES" ...
  .. ..$ : chr [1:5] "E" "A" "C" "ES" ...
 $ Goldberg (1992a)               : num [1:5, 1:5] 1 0.06 0.04 0.16 0.24 0.06 1 0.13 0.23 -0.09 ...
  ..- attr(*, "dimnames")=List of 2
  .. ..$ : chr [1:5] "E" "A" "C" "ES" ...
  .. ..$ : chr [1:5] "E" "A" "C" "ES" ...
}
\details{
The five factors are:
\describe{
  \item{E}{Extraversion}
  \item{A}{Agreeableness}
  \item{C}{Conscientiousness}
  \item{ES}{Emotional Stability}
  \item{I}{Intellect or Openness to Experience}
}
  The sample sizes are in Digman97.n.
}
\source{
Digman, J.M. (1997). Higher-order factors of the Big Five. \emph{Journal of Personality and Social Psychology}, \bold{73}, 1246-1256.
}
\references{
 Cheung, M. W. L., & Chan, W. (2005). Classifying correlation matrices into relatively homogeneous subgroups: A cluster analytic approach. \emph{Educational and Psychological Measurement}, \bold{65}, 954-979.
}
\examples{
data(Digman97)
## maybe str(Digman97) ; plot(Digman97) ...
}
\keyword{datasets}
