\name{Digman97}
\alias{Digman97}
\alias{Digman97.n}
\alias{Digman97.cluster}
\docType{data}
\title{Factor Correlation Matrices of Big Five Model from Digman (1997)}
\description{
Fourteen studies of the factor correlation matrices of the Big Five Model of personality reported by Digman (1997). 
}
\usage{data(Digman97)}
\format{
}
\details{
The five factors are:
\describe{
  \item{E}{Extraversion}
  \item{A}{Agreeableness}
  \item{C}{Conscientiousness}
  \item{ES}{Emotional Stability}
  \item{I}{Intellect or Openness to Experience}
}
  Digman97.n and Digman97.cluster store the sample sizes and the types
  of participants of the studies, respectively.
}
\source{
Digman, J.M. (1997). Higher-order factors of the Big Five. \emph{Journal of Personality and Social Psychology}, \bold{73}, 1246-1256.
}
\references{
 Cheung, M. W. L., & Chan, W. (2005). Classifying correlation matrices into relatively homogeneous subgroups: A cluster analytic approach. \emph{Educational and Psychological Measurement}, \bold{65}, 954-979.
}
\examples{
data(Digman97)
## maybe str(Digman97)  ...
}
\keyword{datasets}
