\name{Digman97}
\alias{Digman97}
\alias{Digman97.n}
\alias{Digman97.cluster}
\docType{data}
\title{Factor Correlation Matrices of Big Five Model from Digman (1997)}
\description{
Fourteen studies of the factor correlation matrices of the Five-Factor Model of personality reported by Digman (1997). 
}
\usage{data(Digman97)}
\format{
}
\details{
A list of data with the following structure:
\describe{
  \item{data}{A list of 14 studies of correlation matrices. The
	variables are Extraversion (E), Agreeableness (A), Conscientiousness
	(C), Emotional Stability (ES) and Intellect (I)}
  \item{n}{A vector of sample sizes}
  \item{cluster}{Types of participants of the studies}
}
}
\source{
Digman, J.M. (1997). Higher-order factors of the Big Five. \emph{Journal of Personality and Social Psychology}, \bold{73}, 1246-1256.
}
\references{
 Cheung, M. W. L., & Chan, W. (2005). Classifying correlation matrices into relatively homogeneous subgroups: A cluster analytic approach. \emph{Educational and Psychological Measurement}, \bold{65}, 954-979.
}
\examples{
data(Digman97)
## maybe str(Digman97)  ...
}
\keyword{datasets}
