\name{issp89}
\alias{issp89}
\docType{data}
\title{Data Set from Cheung and Chan (2005; 2009) 
}
\description{Eleven covariance matrices on work-related attitudes were extracted from the Inter-University Consortium for Political and Social Research (1989). Nine variables were selected by Cheung and Chan (2005; 2009) for demonstration purposes. They were grouped into three meaningful constructs: job prospects (F1), including job security (x1), income (x2), and advancement opportunity (x3); job nature (F2), including interesting job (x4), independent work (x5), help other people (x6), and useful to society (x7); and time demand (F3), represented by flexible working hours (x8) and lots of leisure time (x9).
}
\usage{data(issp89)}
\format{
}
\details{
A list of data with the following structure:
\describe{
  \item{data}{A list of 11 studies of correlation matrices }
  \item{n}{A vector of sample sizes}
}
}
\source{Inter-University Consortium for Political and Social Research. (1989). \emph{International Social Survey Program: Work orientation}. Ann Arbor, MI: Author.}

\references{
  Cheung, M. W. L., & Chan, W. (2005). Meta-analytic structural equation modeling: A two-stage approach. \emph{Psychological Methods}, \bold{10}, 40-64.

Cheung, M. W. L., & Chan, W. (2009). A two-stage approach to synthesizing covariance matrices in meta-analytic structural equation modeling. \emph{Structural Equation Modeling}, \bold{16}, 28-53.
}
\examples{
data(issp89)
## maybe str(issp89)
}
\keyword{datasets}
