\name{Norton13}
\alias{Norton13}
\docType{data}
\title{Studies on the Hospital Anxiety and Depression Scale Reported by Norton et al. (2013)
}
\description{
  The data set includes 28 studies on 14 items measuring the
  Hospital Anxiety and Depression Scale (HADS) Reported by Norton et
  al. (2013).
}
\usage{data(Norton13)}

\details{
      The variables are:
  \describe{
  \item{data}{A list of 28 studies of correlation matrices. The
      variables are 14 items (x1 to x14) measuring HADS.}
  \item{n}{A vector of sample sizes}
  \item{population}{A vector of the population of the data}
  \item{group}{A vector of classification into \emph{patients}
      vs. \emph{non-patients} based on population}
  }
}
\source{
Norton, S., Cosco, T., Doyle, F., Done, J., & Sacker, A. (2013). The Hospital Anxiety and Depression Scale: A meta confirmatory factor analysis. \emph{Journal of Psychosomatic Research}, \emph{74}(1), 74-81.
}
\references{
  Jak, S., & Cheung, M. W.-L. (in press). Addressing heterogeneity in meta-analytic structural equation modeling using subgroup analysis. \emph{Behavior Research
Methods}.
}
\examples{
data(Norton13)
}
\keyword{datasets}
